confirmation dialog when deleting a label

+++ frontend/src/main/kotlin/views/SettingsLabelView.kt
@@ -1,10 +1,14 @@
 package views
 
 import controllers.GroupViewController
+import javafx.scene.control.Alert
+import javafx.scene.control.Alert.AlertType
+import javafx.scene.control.ButtonType
 import javafx.scene.control.ColorPicker
 import javafx.scene.layout.HBox
 import javafx.scene.paint.Color
 import models.Label
+import java.util.*
 
 class SettingsLabelView(
     private val groupController: GroupViewController,
@@ -83,6 +87,15 @@ class SettingsLabelView(
     }
 
     override fun deleteLabel() {
-        groupController.deleteLabel(label)
+        val alert = Alert(AlertType.CONFIRMATION)
+        alert.title = "Confirmation Dialog"
+        alert.headerText = "Are you sure you want to delete this label?"
+
+        val result: Optional<ButtonType> = alert.showAndWait()
+        if (result.get() === ButtonType.OK) {
+            groupController.deleteLabel(label)
+        } else {
+            // do nothing
+        }
     }
 }
